api/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.30.0
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Storage
boto3==1.28.57

# HTTP Client
httpx==0.25.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.0

# OpenAI
openai==1.3.0

# Background Jobs
redis==5.0.1
rq==1.15.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2024.1

# CORS is built into FastAPI
